% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbum.R
\docType{package}
\name{bbum}
\alias{bbum}
\title{bbum: A package for correcting for secondary effects signals in multiple testing}
\description{
Standard multiple testing correction methods
cannot distinguish dominant primary effects from
background secondary effects above null. A bi-beta-uniform
mixture (BBUM) model allows the modeling and correction
for the false discovery rate (FDR) of both null and
secondary effects.
}
\section{Authors}{

\itemize{
\item Author: Peter Y. Wang
\item Advisor: David P. Bartel
}
}

\references{
Wang PY, Bartel DP. 2021. TBD. \emph{bioRxiv}. doi:TBD

Allison DB, Gadbury GL, Heo M, Fernández JR, Lee C-K, Prolla TA, Weindruch R.
2002. A mixture model approach for the analysis of microarray gene
expression data. \emph{Computational Statistics & Data Analysis} 39:1-20.
doi:10.1016/S0167-9473(01)00046-9

Markitsis A, Lai Y. 2010. A censored beta mixture model for the estimation of
the proportion of non-differentially expressed genes. \emph{Bioinformatics}
26:640-646. doi:10.1093/bioinformatics/btq001

Pounds S, Morris SW. 2003. Estimating the occurrence of false positives and
false negatives in microarray studies by approximating and partitioning the
empirical distribution of p-values. \emph{Bioinformatics} 19:1236-1242.
doi:10.1093/bioinformatics/btg148
}
