% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBUM_DEcorr.R
\name{BBUM_DEcorr}
\alias{BBUM_DEcorr}
\title{FDR correction of secondary effects and significance calling after DESeq2}
\usage{
BBUM_DEcorr(
  df.deseq,
  classCol,
  geneName = NULL,
  pBBUM.alpha = 0.05,
  excluded = c(),
  outliers = c(),
  add_starts = list(),
  only_start = FALSE,
  limits = list(),
  auto_outliers = TRUE,
  rthres = 1,
  rtrimmax = 0.05,
  atrimmax = 10,
  quiet = FALSE
)
}
\arguments{
\item{df.deseq}{The output DESeqResults object from DESeq2, after running
results(). This object should normally be of \code{data.frame} subclass.
The \code{data.frame} could also have been further processed before calling
\code{BBUM_DEcorr}, as long as the \code{pvalue} column is unchanged.}

\item{classCol}{A \code{str} of the name of the column that indicates whether
a data point belongs in the signal set or the background set.}

\item{geneName}{A \code{str} of the name of the column that contains the
names of genes, e.g. \code{"FeatureID"}. Leave as default (\code{NULL}) if
the column has not been changed (as row names, or \code{"row"} if
\code{tidy=T} in \code{results()}).}

\item{pBBUM.alpha}{Cutoff level of \code{pBBUM} for significance calling.}

\item{excluded}{A \code{str} vector of all gene names that did \strong{not} meet
some user-determined prior filtering criteria (e.g. read cutoff). If
\code{excluded} is not supplied, all genes with valid p values will be used
for analysis.}

\item{outliers}{A \code{str} vector of all (additional) genes names that
should be regarded as outliers among the background set.}

\item{add_starts}{List of named vectors for additional starts of fitting
algorithm.}

\item{only_start}{Whether the algorithm should only use the given starts to fit.}

\item{limits}{Named list of custom limits for specific paramters. Parameters
not mentioned would be default values.}

\item{auto_outliers}{Toggle automatic outlier trimming.}

\item{rthres}{Threshold value of \code{r} parameter to trigger a failed
\code{r.pass} value.}

\item{rtrimmax}{Maximum fraction of data points allowed to be outliers in the
background set of data (to be trimmed).}

\item{atrimmax}{Maximum absolute number of data points allowed to be outliers
in the background set of data (to be trimmed).}

\item{quiet}{Suppress printed messages and warnings.}
}
\value{
A \code{data.frame} based on the input results table, with added
columns from BBUM correction and significance calling:
\itemize{
\item \code{geneName}: Copy of extracted gene names (\code{str})
\item \code{excluded}: Whether the point is considered failing the cutoff and
excluded in \code{excluded} (\code{bool})
\item \code{outlier}: Whether the point is considered an outlier (\code{bool})
\item \code{BBUM.class}: Whether the data point is considered in the singal set
instead of the background set (\code{bool})
\item \code{BBUM.l}, \code{BBUM.a}, \code{BBUM.th}, \code{BBUM.r}: Coefficients
from BBUM model fit (\code{float})
\item \code{pBBUM}: BBUM-transformed, FDR-adusted p value (\code{float})
\item \code{BBUM.hit}: Whether a gene is a hit after signficance calling
(\code{bool})
\item \code{BBUM.fct}: Factor summarizing status as hits, non-hits, or outliers
(\code{factor})
}
}
\description{
Process DESeq2 data tables. Out of two subsets of data rows, one with
primary effects and one without, \code{BBUM_DEcorr} models the p values to
distinguish secondary effects signal from primary effects signal using the
BBUM model. It then calls significant genes within the signal set after
correcting for FDR of \strong{both} null \strong{and} secondary effects in multiple
testing.
}
\details{
The output \code{data.frame} preserves all original rows in order,
all original columns in order, and row names if present.

\code{classCol}: For example, in the case where positive log fold
changes represent the signal class, and negative log fold changes represent
the background class, all positive fold change points have
\code{classCol == T}, and all negative fold change points have
\code{classCol == F}.

\code{pBBUM} represents the expected overall FDR level if the cutoff
were set at that particular p value. This is similar to the interpretation
of p values corrected through the typical \code{p.adjust(method = "fdr")}.

\code{BBUM_DEcorr} functions properly only with p values filtered
for poor quality data points, e.g. low read counts. Filtering has either
been done manually in prior, through DESeq2's \code{independentFiltering}
in \code{results()}, or through the \code{excluded} argument. Excluding a
point via \code{excluded} is equivalent to filtering it out of the
data.frame to begin with.

Outliers (\code{outliers}) are included in the graphs from
\code{BBUM_plot()} and have p values associated. Excluded points
(\code{excluded}) are removed from analysis altogether.

Default starts for BBUM fitting are implemented. If additional
starts should included, or only custom starts should be considered, make
use of \code{add_starts} and/or \code{only_start} arguments.

If more than one start achieved the identical maximum likelihood,
A random start is chosen among them.

Automatic outlier detection relies on the model fitting a value of
\code{r > 1}. For benchmarking of the trimming strategy, see
Wang & Bartel, 2021.

Adding too many starts or allowing too much outlier trimming can
increase computation time.
}
\examples{
\dontrun{
# Typical default run
dds = DESeqDataSetFromMatrix(countData = cts, ...) # DESeq2
dds = DESeq(dds)
res = results(dds) \%>\%
  as.data.frame() \%>\%
  mutate(FCup = log2FoldChange > 0)
res.BBUMcorr = BBUM_DEcorr(
  deseq_results = res,
  classCol = "FCup"
  ) # function call

# Some customized behavior
res.BBUMcorr =  BBUM_DEcorr(
  deseq_results = res,
  classCol = "signal",
  geneName = "miRNAs",
  pBBUM.alpha = 0.01,
  excluded = c("hsa-miR-124-5p", "hsa-miR-1-3p"),
  outliers = c("hsa-miR-21-5p"),
  add_starts = list(c(lambda = 0.9, a = 0.6, theta = 0.1, r = 0.1))
  )
}

}
