% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBUM_params.R
\name{BBUM_default_lim}
\alias{BBUM_default_lim}
\title{Default static limits for BBUM fitting}
\usage{
BBUM_default_lim(rcap)
}
\arguments{
\item{rcap}{Whether the parameter r should have a stringent upper bound in
this instance.}
}
\description{
For internal use of the package.
}
