% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBUM_distribution.R
\name{BBUM_distribution}
\alias{BBUM_distribution}
\alias{dbbum}
\alias{pbbum}
\alias{rbbum}
\alias{rbbum.ID}
\title{Bi-beta-uniform mixture (BBUM) distribution}
\usage{
dbbum(x, lambda, a, theta, r)

pbbum(q, lambda, a, theta, r)

rbbum(n, lambda, a, theta, r)

rbbum.ID(n, lambda, a, theta, r)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{lambda}{Vector of BBUM parameter \code{lambda}. lambda is the fraction
of null (uniform distribution density) over all density except the primary
beta distribution.}

\item{a}{Vector of BBUM parameter \code{a}, which corresponds to the a
parameter of the secondary beta distribution component. It describes the
steepness of the second beta distribution.}

\item{theta}{Vector of BBUM parameter \code{theta}. theta is the fraction of
primary beta distribution density over all density.}

\item{r}{Vector of BBUM parameter \code{r}, which is the ratio of the a
parameter of the primary beta distribution over that of the secondary beta
distribution. In other words, \code{a*r} is the a parameter for the primary
beta distribution component.}

\item{n}{Number of observations.}
}
\value{
\code{dbbum} gives the density (PDF), and \code{pbbum} gives the
distribution function (CDF). \code{rbbum} generates random values from the
distribution, and \code{rbbum.ID} generates random values while identifying
the part of BBUM they belong to, as a \code{data.frame} of two columns with
length \code{n}. \code{pvalue} corresponds to the generated values, while
\code{cate} is \code{0} for null, \code{1} for primary beta, and \code{2}
for secondary beta.
}
\description{
Density, distribution function, and random generation for the BBUM
distribution, with four parameters, \code{lambda}, \code{a}, \code{theta},
and \code{r}. The quantile function cannot be simply expressed and is thus
not available.
}
\details{
The \code{b} parameter of the typical beta distribution function is
constrained to be 1 here to limit it to a one-sided, monotonic case,
peaking at 0, when 0 < \code{a} < 1.
}
\examples{
dbbum(x = 0.096, lambda = 0.65, a = 0.1, theta = 0.02, r = 0.07)
pbbum(q = 0.096, lambda = 0.65, a = 0.1, theta = 0.02, r = 0.07)

dbbum(x = c(0.013, 0.04, 0.93, 0.8), lambda = 0.73, a = 0.02, theta = 0.11, r = 0.003)
pbbum(q = c(0.013, 0.04, 0.93, 0.8), lambda = 0.73, a = 0.02, theta = 0.11, r = 0.003)

rbbum(n = 100, lambda = 0.65, a = 0.1, theta = 0.02, r = 0.07)

rbbum.ID(n = 100, lambda = 0.65, a = 0.1, theta = 0.02, r = 0.07)

}
