% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBUM_expectedhits.R
\name{BBUM_expectedhits}
\alias{BBUM_expectedhits}
\title{Expected hits count under the BBUM model}
\usage{
BBUM_expectedhits(params, alpha, n)
}
\arguments{
\item{params}{Named list of BBUM parameters.}

\item{alpha}{The cutoff selected for BBUM-FDR-corrected p values.}

\item{n}{Total number of considered data points. This should only include the
signal set.}
}
\value{
The numeric value of the expected hits count, as a decimal numeric.
}
\description{
\code{BBUM_expectedhits} computes the expected number of true hits, according
to the given BBUM distribution and the pBBUM value cutoff (\code{alpha})
selected.
}
\details{
\code{params} should be named as \code{BBUM.l}, \code{BBUM.a},
\code{BBUM.th}, \code{BBUM.a}.
}
\examples{
BBUM_expectedhits(
  params = list(BBUM.l = 0.65, BBUM.a = 0.1, BBUM.th = 0.02, BBUM.r = 0.07),
  alpha = 0.05,
  n = 432
  )

}
