% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBUM_fit.R
\name{BBUM_fit}
\alias{BBUM_fit}
\title{BBUM statistical modeling}
\usage{
BBUM_fit(
  dt_signal_set = NULL,
  dt_bg_set = NULL,
  dt_all = NULL,
  signal_set = NULL,
  starts,
  limits = list(),
  rcap = TRUE,
  outlier_trim = 0,
  rthres = 1
)
}
\arguments{
\item{dt_signal_set, dt_bg_set}{Vectors of numerical p values, belonging to
the signal set and background set respectively.}

\item{dt_all}{A vector of all numerical p values, including both signal and
background sets.}

\item{signal_set}{A vector of booleans signifying which values among
\code{dt_all} are signal set data points. Used only in conjunction with
\code{dt_all}.}

\item{starts}{A list of named vectors of starts for the four BBUM parameters.}

\item{limits}{Named list of custom limits for specific paramters. Parameters
not mentioned would be default values.}

\item{rcap}{Whether the parameter r should have a stringent upper bound in
this instance.}

\item{outlier_trim}{Number of strongest points among the background class to be
trimmed as outliers.}

\item{rthres}{Threshold value of \code{r} parameter to trigger a failed
\code{r.pass} value.}
}
\value{
A named list with the following items:
\itemize{
\item \code{LL}: Value of the maximized log-likelihood.
\item \code{val.l}: Fitted \code{lambda} parameter.
\item \code{val.a}: Fitted \code{a} parameter.
\item \code{val.th}: Fitted \code{theta} parameter.
\item \code{val.r}: Fitted \code{r} parameter.
\item \code{fit.obj}: Fit output object of \code{optim}.
}
}
\description{
Fitting the BBUM model on data containing a set with primary effects (signal
set) and a set without (background set), using maximum likelihood estimation
(MLE) with the Nelder-Mead algorithm. It chooses the best solution among
all starts provided.
}
\details{
Either use \code{dt_signal_set} and \code{dt_bg_set} to input data
separately, or use \code{dt_all} and \code{signal_set} to input data.
When both pairs are defined, \code{dt_all} and \code{signal_set} are
ignored.

If more than one start achieved the identical maximum likelihood,
A random start is chosen among them.

\code{rcap} is used internally to decide on the default limits for
\code{r}.

A failed \code{r.pass} code is not
triggered if \code{lambda} is too big for a reliable fitting of \code{a} to
begin with.
}
\examples{
BBUM_fit(
  dt_signal_set = c(0.031, 0.0014, 0.0003, 0.0001, 0.049,
                    0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9),
  dt_bg_set     = c(0.701, 0.503, 0.109, 0.0071, 0.019,
                    0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9),
  starts = list(c(lambda = 0.9, a = 0.1, theta = 0.1, r = 0.1))
)
BBUM_fit(
  dt_all        = c(0.701, 0.503, 0.109, 0.0071, 0.019, 0.031, 0.0014,
                    0.0003, 0.0001, 0.049, 0.0001,
                    0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9,
                    0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9),
  signal_set    = c(FALSE, FALSE, FALSE, FALSE, FALSE, TRUE, TRUE,
                    TRUE, TRUE, TRUE, FALSE,
                    FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,
                    TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE),
  starts = list(c(lambda = 0.9, a = 0.1, theta = 0.1, r = 0.1)),
  outlier_trim = 1
)

}
