% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBUM_loglik.R
\name{BBUM_loglik}
\alias{BBUM_loglik}
\title{Log-likelihood for the BBUM model}
\usage{
BBUM_loglik(params, posSet, negSet, limits = list(), rcap)
}
\arguments{
\item{params}{Named vector of BBUM parameters.}

\item{posSet}{Vector of values following a BBUM distribution including the
primary beta component.}

\item{negSet}{Vector of values following a BBUM distribution without the
primary beta component, i.e. a BUM distribution with the secondary beta
distribution.}

\item{limits}{Named list of custom limits for specific paramters. Parameters
not mentioned would be default values.}

\item{rcap}{Whether the parameter r should have a stringent upper bound in
this instance.}
}
\value{
The value of the total log-likelihood. The logarithm used is the
natural logarithm.
}
\description{
\code{BBUM_loglik} computes the total log-likelihood of the given BBUM
parameters, given data points with and without the primary beta component.
This is the maximization target for fitting the BBUM model to data.
}
