% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBUM_plot.R
\name{BBUM_plot}
\alias{BBUM_plot}
\title{QC and results graphs for the BBUM-processed results data frame}
\usage{
BBUM_plot(
  df.bbum,
  option = c("MA", "volcano", "hist", "ecdf", "ecdf_log", "ecdf.corr", "ecdf_log.corr",
    "pp", "pcorr", "symm"),
  expressionCol = "baseMean",
  pBBUM.alpha = 0.05
)
}
\arguments{
\item{df.bbum}{The data.frame output of \code{BBUM_DEcorr()}, with any
additional columns.}

\item{option}{Option of graph to plot. If a vector of length > 1 is provided,
only the first element is used. Ignores case.}

\item{expressionCol}{A \code{str} of the name of the column used for the x
axis of the MA graph. (Only for \code{option = "MA"}.)}

\item{pBBUM.alpha}{Cutoff level of \code{pBBUM} for significance calling.}
}
\value{
\code{ggplot2} plot object.
}
\description{
Useful graphs for checking and viewing the results of BBUM correction and
significance calling.
}
\details{
The argument \code{expressionCol} allows plotting the MA graph
against a specified column as the x axis (expression level). For instance,
some may prefer to plot the mean normalized expression in control
experiments only, rather than the default \code{"baseMean"} of DESeq2.

Graph \code{option}s are:
\itemize{
\item \code{MA}: MA plot (\code{log2FoldChange} against \code{expressionCol})
\item \code{volcano}: Volcano plot (\code{-log10(pvalue)} against
\code{log2FoldChange})
\item \code{hist}: p value histogram separated into signal and background set
points, with BBUM model overlaid. Background histogram is normalized by a
factor of \code{1 - theta} to account for the lack of primary effects for
comparison.
\item \code{ecdf}: p value ECDF separated into signal and background set
points, with BBUM model overlaid.
\item \code{ecdf_log}: ECDF in log scale for the p values, which helps to focus
on the left-tail.
\item \code{ecdf.corr}, \code{ecdf_log.corr}: Plots the \code{pBBUM} values
instead to evaluate the FDR-corrected p values.
\item \code{pp}: P-P plot to evaluate the goodness of fit.
\item \code{pcorr}: Plot of p values from raw values to BBUM-FDR_corrected
values, by data set. This plot is helpful for evaluating the correction of
individual p values through the BBUM algorithm.
\item \code{symm}: Modified symmetry plot of \code{-log10(p)} values, excluding
hits, with \code{-log10(0)} as the mid-point instead of the median. Uses
subsampling to account for the different number of points in the signal and
background sets.
}

The most critical region of BBUM distribution for an appropriate
correction for secondary effects is the "left-tail" around 0, where both
primary and secondary beta components peak. An ECDF graph in log scale
allows emphasis and better visualization of this region.

ECDF graphs are overlaid on the \code{x = y} diagonal line, which
represents the uniform/null-only i.e. no secondary effects case.

Because the peak near \code{p = 0} is the most informative region
for p values correction, a P-P plot is more appropriate to assess
goodness-of-fit of BBUM models than a Q-Q plot.

Plot \code{symm} is for the validation of the assumption that the
signal and background sets have roughly similar background (null and
secondary effects) distributions of p values. As excluding hits does not
exclude the false-negative region, there is still an expected discrepancy
at low p values. The implemented color gradient attempts to reflect
this expected up-deviation from the diagonal line when the fraction of
remaining primary effects is large. Empirically, distributions that do not
deviate from the \code{+/- log(10)} dashed lines when the expected primary
effects fractions is low are symmetrical enough for accurate BBUM
correction.
}
\examples{
\dontrun{
BBUM_plot(df.bbum = res.BBUMcorr,
          option = "ecdf_log",
          expressionCol = "WTmean",
          pBBUM.alpha = 0.01)
}

}
