% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBUM_startgen.R
\name{BBUM_startgen}
\alias{BBUM_startgen}
\title{Generating guess values}
\usage{
BBUM_startgen(limits = list())
}
\arguments{
\item{limits}{Named list of custom limits for specific paramters. Parameters
not mentioned would be default values.}
}
\description{
Generates properly rescaled start values for BBUM fitting, including 6 static
configurations by default.
}
