% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_rescale.R
\name{lin_rescale}
\alias{lin_rescale}
\title{Linear rescaling}
\usage{
lin_rescale(x, left, right)
}
\arguments{
\item{x}{A numeric vector.}

\item{left, right}{The lower and upper boundaries for rescaled interval.}
}
\value{
A vector with the same length as \code{x}.
}
\description{
Rescale numeric values from 0 to 1 to a specified interval.
}
\details{
Values smaller than 0 or larger than 1 are rescaled to values
linearly extrapolated from the specified interval.
}
\examples{
lin_rescale(0.4, left = 100, right = 200)
lin_rescale(c(-2,0.3,0.6,10), left = -10, right = 10)

}
