% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\alias{inv.logit}
\title{Logit function}
\usage{
logit(x, base = exp(1), left = 0, right = 1, suppress.nan = TRUE)

inv.logit(x, base = exp(1), left = 0, right = 1)
}
\arguments{
\item{x}{A numeric vector.}

\item{base}{The base used for the logarithm.}

\item{left, right}{The lower and upper domain boundaries of the corresponding
logit function. Defaults are 0 and 1.}

\item{suppress.nan}{Whether to output NaN or -Inf/Inf when x is out of
bounds. If TRUE, an out-of-bounds warning is issued while suppressing the
NaN into either Inf (x > right) or -Inf (x < left).}
}
\value{
A vector with the same length as \code{x}.
}
\description{
\code{logit} computes values transformed by the logit or log-odds function.
\code{inv.logit} is the inverse of \code{logit}. The logarithm base used
can be defined (default is natural logarithm). The left and right bounds
of the logit function can also be defined, with default domain being
\code{(0, 1)}.
}
\examples{
logit(0.7)
logit(6.2, left = 0, right = 10)
logit(-0.07, suppress.nan = FALSE)
logit(0.01, base = 10)

inv.logit(2.31)
inv.logit(5.1, left = -1, right = 1)
inv.logit(2, base = 2)

}
