% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBUM_distribution.R
\name{qbeta_a}
\alias{qbeta_a}
\title{Quantile function of beta distribution with b = 1}
\usage{
qbeta_a(p, a)
}
\arguments{
\item{p}{Vector of probabilities.}

\item{a}{Vector of a / shape1.}
}
\value{
The quantile function.
}
\description{
Quantile function for the special case of the beta distribution, where b = 1.
}
\examples{
qbeta_a(p = 0.013, a = 0.1)

}
