% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbum.R
\docType{package}
\name{bbum}
\alias{bbum}
\title{bbum: A package for correcting for secondary effects signals in multiple testing}
\description{
Standard multiple testing correction methods cannot directly handle datasets
that contain a background secondary signal confounding the primary signal
of interest. A bi-beta-uniform mixture (BBUM) model allows the modeling
and correction for the false discovery rate (FDR) of both null and
secondary effects.
}
\section{Authors}{

\itemize{
\item Author: Peter Y. Wang
\item Advisor: David P. Bartel
}
}

\references{
Wang PY, Bartel DP. 2022. TBD. \emph{bioRxiv}. doi:TBD

Markitsis A, Lai Y. 2010. A censored beta mixture model for the estimation of
the proportion of non-differentially expressed genes. \emph{Bioinformatics}
\strong{26}:640-646. doi:10.1093/bioinformatics/btq001

Pounds S, Morris SW. 2003. Estimating the occurrence of false positives and
false negatives in microarray studies by approximating and partitioning the
empirical distribution of p-values. \emph{Bioinformatics} \strong{19}:1236-1242.
doi:10.1093/bioinformatics/btg148
}
