% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbum.R
\docType{package}
\name{bbum}
\alias{bbum}
\title{bbum: A package for correcting secondary effects signals in multiple testing}
\description{
Standard multiple testing correction methods
cannot distinguish dominant primary effects from
background secondary effects above null. A bi-beta-uniform
mixture (BBUM) model allows the modeling and correction
for the false discovery rate (FDR) of both null and
secondary effects.
}
\section{Authors}{

\itemize{
\item Author: Peter Y. Wang
\item Advisor: David P. Bartel
}
}

\references{
TODO
}
