% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBUM_apply.R
\name{BBUM_apply}
\alias{BBUM_apply}
\title{Apply BBUM fitting and correction onto a data.frame}
\usage{
BBUM_apply(
  df,
  add_starts,
  only_start,
  limits,
  auto_outliers,
  rthres,
  rtrimmax,
  atrimmax,
  quiet
)
}
\arguments{
\item{df}{\code{data.frame} input}

\item{add_starts}{List of named vectors for additional starts of fitting
algorithm.}

\item{only_start}{Whether the algorithm should only use the given starts to fit.}

\item{limits}{Named list of custom limits for specific paramters. Parameters
not mentioned would be default values.}

\item{auto_outliers}{Toggle automatic outlier trimming.}

\item{rthres}{Threshold value of \code{r} parameter to trigger a failed
\code{r.pass} value.}

\item{rtrimmax}{Maximum fraction of data points allowed to be outliers in the
background set of data (to be trimmed).}

\item{atrimmax}{Maximum absolute number of data points allowed to be outliers
in the background set of data (to be trimmed).}

\item{quiet}{Suppress printed messages and warnings.}
}
\description{
Wrapper/intermediate function between \code{BBUM_corr} and
\code{BBUM_DEcorr}. For internal use of the package.
}
