% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBUM_corr.R
\name{BBUM_corr}
\alias{BBUM_corr}
\title{BBUM FDR correction of p values}
\usage{
BBUM_corr(
  pvals,
  signal_set,
  add_starts = list(),
  only_start = FALSE,
  limits = list(),
  auto_outliers = TRUE,
  rthres = 1,
  rtrimmax = 0.05,
  atrimmax = 10,
  quiet = FALSE
)
}
\arguments{
\item{pvals}{A vector of all numerical p values, including both signal and
background sets.}

\item{signal_set}{A vector of booleans signifying which values among
\code{dt_all} are signal set data points. Used only in conjunction with
\code{dt_all}.}

\item{add_starts}{List of named vectors for additional starts of fitting
algorithm.}

\item{only_start}{Whether the algorithm should only use the given starts to fit.}

\item{limits}{Named list of custom limits for specific paramters. Parameters
not mentioned would be default values.}

\item{auto_outliers}{Toggle automatic outlier trimming.}

\item{rthres}{Threshold value of \code{r} parameter to trigger a failed
\code{r.pass} value.}

\item{rtrimmax}{Maximum fraction of data points allowed to be outliers in the
background set of data (to be trimmed).}

\item{atrimmax}{Maximum absolute number of data points allowed to be outliers
in the background set of data (to be trimmed).}

\item{quiet}{Suppress printed messages and warnings.}
}
\value{
A named list with the following items:
\itemize{
\item \code{pvals}: Vector of input p values.
\item \code{pBBUMs}: Vector of p values corrected for FDR by BBUM modeling.
\item \code{BBUM.l}: Fitted \code{lambda} parameter.
\item \code{BBUM.a}: Fitted \code{a} parameter.
\item \code{BBUM.th}: Fitted \code{theta} parameter.
\item \code{BBUM.r}: Fitted \code{r} parameter.
\item \code{outlier_trim}: Number of outliers trimmed in the background set.
\item \code{r.passed}: Boolean for whether the fitted \code{r} value was under
the threshold for flagging outliers.
\item \code{BBUM.LL}: Value of the maximized log-likelihood.
\item \code{fitdist_obj}: Fit output object of \code{optim}.
}
}
\description{
Fits the BBUM model on the data, and transforms raw p values into values
corrected for false discovery rate (FDR) of null \strong{and} secondary effects
according to the BBUM model, as multiple testing correction. Optionally, it
automatically detects extreme value outliers among the background set, and
resolves correction issues by trimming the outliers.
}
\details{
\code{pBBUM} represents the expected overall FDR level if the cutoff
were set at that particular p value. This is similar to the interpretation
of p values corrected through the typical \code{p.adjust(method = "fdr")}.

Default starts for BBUM fitting are implemented. If additional
starts should included, or only custom starts should be considered, make
use of \code{add_starts} and/or \code{only_start} arguments.

If more than one start achieved the identical maximum likelihood,
A random start is chosen among them.

Automatic outlier detection relies on the model fitting a value of
\code{r > 1}. For benchmarking of the trimming strategy, see
Wang & Bartel, 2021.

Adding too many starts or allowing too much outlier trimming can
increase computation time.
}
\examples{
BBUM_corr(
  pvals        = c(0.701, 0.503, 0.109, 0.0071, 0.019, 0.031, 0.0014,
                    0.0003, 0.0001, 0.049, 0.0001,
                    0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9,
                    0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9),
  signal_set    = c(FALSE, FALSE, FALSE, FALSE, FALSE, TRUE, TRUE,
                    TRUE, TRUE, TRUE, FALSE,
                    FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,
                    TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE),
  add_starts = list(c(lambda = 0.9, a = 0.6, theta = 0.1, r = 0.1)),
  limits = list(a = c(0.50,0.75))
)

}
