% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBUM_default_lim.R
\name{BBUM_default_lim}
\alias{BBUM_default_lim}
\title{Default static limits for BBUM fitting}
\usage{
BBUM_default_lim(rcap, pBBUM.alpha)
}
\arguments{
\item{rcap}{Whether the parameter r should have a stringent upper bound in
this instance (for smart toggling of outlier detection).}

\item{pBBUM.alpha}{Cutoff level of BBUM-FDR-adjusted p values for
significance testing. Only used here to generate appropriate default
limits.}
}
\description{
For internal use of the package.
}
