% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBUM_fit.R
\name{BBUM_fit}
\alias{BBUM_fit}
\title{BBUM statistical modeling}
\usage{
BBUM_fit(
  dt_signal_set = NULL,
  dt_bg_set = NULL,
  dt_all = NULL,
  signal_set = NULL,
  starts,
  limits = list(),
  rcap = TRUE,
  outlier_trim = 0,
  rthres = 1
)
}
\arguments{
\item{dt_signal_set, dt_bg_set}{Vectors of numerical p values, belonging to
the signal set and background set respectively.}

\item{dt_all}{A vector of all numerical p values, including both signal and
background sets.}

\item{signal_set}{A vector of booleans signifying which values among
\code{dt_all} are signal set data points. Used only in conjunction with
\code{dt_all}.}

\item{starts}{A list of named vectors of starts for the four BBUM parameters.}

\item{limits}{Named list of custom limits for specific paramters. Parameters
not mentioned would be default values.}

\item{rcap}{Whether the parameter r should have a stringent upper bound in
this instance.}

\item{outlier_trim}{Number of strongest points among the background class to be
trimmed as outliers.}

\item{rthres}{Threshold value of \code{r} parameter to trigger a failed
\code{r.pass} value.}
}
\value{
A named list with the following items:
\itemize{
\item \code{estim}: A named list of fitted parameter values.
\item \code{LL}: Value of the maximized log-likelihood.
\item \code{convergence}: Convergence code from \code{optim}.
\item \code{outlier_trim}: The input value of the \code{outlier_trim} argument.
\item \code{r.passed}: Boolean for whether the fitted \code{r} value was under
the threshold for flagging outliers.
}
}
\description{
Fitting the BBUM model on data containing a set with primary effects (signal
set) and a set without (background set), using maximum likelihood estimation
(MLE) with the Nelder-Mead algorithm. It chooses the best solution among
all starts provided.
}
\details{
Either use \code{dt_signal_set} and \code{dt_bg_set} to input data
separately, or use \code{dt_all} and \code{signal_set} to input data.
When both pairs are defined, \code{dt_all} and \code{signal_set} are
ignored.

If more than one start achieved the identical maximum likelihood,
A random start is chosen among them.

\code{rcap} is used internally to decide on the default limits for
\code{r}.

A failed \code{r.pass} code is not
triggered if \code{lambda} is too big for a reliable fitting of \code{a} to
begin with.
}
\examples{
BBUM_fit(
  dt_signal_set = c(0.000021, 0.00010, 0.03910, 0.031, 0.001,
                    0.13, 0.21, 0.38, 0.42, 0.52, 0.60, 0.73, 0.81, 0.97),
  dt_bg_set     = c(0.501, 0.203, 0.109, 0.071, 0.019,
                    0.11, 0.27, 0.36, 0.43, 0.50, 0.61, 0.77, 0.87, 0.91),
  starts = list(c(lambda = 0.9, a = 0.6, theta = 0.1, r = 0.1))
)
BBUM_fit(
  dt_all        = c(0.501, 0.203, 0.109, 0.071, 0.019, 0.031, 0.001,
                    0.000021, 0.00010, 0.03910,
                    0.0001,
                    0.11, 0.27, 0.36, 0.43, 0.50, 0.61, 0.77, 0.87, 0.91,
                    0.13, 0.21, 0.38, 0.42, 0.52, 0.60, 0.73, 0.81, 0.97),
  signal_set    = c(FALSE, FALSE, FALSE, FALSE, FALSE, TRUE, TRUE,
                    TRUE,  TRUE,  TRUE,
                    FALSE,
                    FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,
                    TRUE,  TRUE,  TRUE,  TRUE,  TRUE,  TRUE,  TRUE,  TRUE,  TRUE),
  starts = list(c(lambda = 0.9, a = 0.6, theta = 0.1, r = 0.1)),
  outlier_trim = 1
)

}
