% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBUM_startgen.R
\name{BBUM_limapply}
\alias{BBUM_limapply}
\title{Rescaling guess values}
\usage{
BBUM_limapply(props, limits = list())
}
\arguments{
\item{props}{Named list of parameter guess as proportions defined in
\code{[0, 1]}.}

\item{limits}{Named list of custom limits for specific paramters. Parameters
not mentioned would be default values.}
}
\description{
Transformer of guess values defined in \code{[0, 1]} to whatever the limits
are.
}
