% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBUM_params.R
\name{BBUM_params_recoverLin}
\alias{BBUM_params_recoverLin}
\title{Transform fitting values to actual values}
\usage{
BBUM_params_recoverLin(params, limits = list(), rcap, pBBUM.alpha)
}
\arguments{
\item{params}{Named vector of BBUM parameters.}

\item{limits}{Named list of custom limits for specific paramters. Parameters
not mentioned would be default values.}

\item{rcap}{Whether the parameter r should have a stringent upper bound in
this instance (for smart toggling of outlier detection).}

\item{pBBUM.alpha}{Cutoff level of BBUM-FDR-adjusted p values for
significance testing. Only used here to generate appropriate default
limits.}
}
\description{
For internal use of the package.
}
