% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBUM_params.R
\name{BBUM_params_toReal}
\alias{BBUM_params_toReal}
\title{Transform actual values to fitting values}
\usage{
BBUM_params_toReal(params, limits = list(), rcap)
}
\arguments{
\item{params}{Named vector of BBUM parameters.}

\item{limits}{Named list of custom limits for specific paramters. Parameters
not mentioned would be default values.}

\item{rcap}{Whether the parameter r should have a stringent upper bound in
this instance.}
}
\description{
For internal use of the package.
}
